Update username/hashtag regex + text replacement
The username and hashtag regexes were updated to support (as close
as possible to) matching valid instagram usernames/hashtags.

Instead of using 'match', and then replacing matching strings in
the main text, regex replacement is used directly. This fixes a
bug where a matching string exists more than once in the main text.